+++ pkg/plugins/gears/types.go
@@ -124,7 +124,7 @@ type Sender struct {
 	errorVelocityCounter          int
 	currentSuccessVelocityCounter int
 	currentErrorVelocityCounter   int
-	currentSec                    int
+	currentSec                    int64
 }
 
 type ManualHashPartitioner struct {
