+++ pkg/plugins/http/types.go
@@ -102,7 +102,7 @@ type Receiver struct {
 	errorVelocityCounter          int
 	currentSuccessVelocityCounter int
 	currentErrorVelocityCounter   int
-	currentSec                    int
+	currentSec                    int64
 }
 
 type SenderConfig struct {
@@ -129,7 +129,7 @@ type Sender struct {
 	errorVelocityCounter          int
 	currentSuccessVelocityCounter int
 	currentErrorVelocityCounter   int
-	currentSec                    int
+	currentSec                    int64
 }
 
 type BadHttpStatusError struct {
